moved module-site-overview-placeholder.jsx to stats-overview-placeholder/index.jsx
