+++ client/my-sites/stats/stats-overview-placeholder/index.jsx
