+++ client/my-sites/stats/stats-overview.jsx
@@ -11,7 +11,7 @@ var React = require( 'react' ),
 var observe = require( 'lib/mixins/data-observe' ),
 	SidebarNavigation = require( 'my-sites/sidebar-navigation' ),
 	SiteOverview = require( './overview' ),
-	SiteOverviewPlaceholder = require( './module-site-overview-placeholder' ),
+	SiteOverviewPlaceholder = require( './stats-overview-placeholder' ),
 	DatePicker = require( './stats-date-picker' ),
 	StatsNavigation = require( './stats-navigation' ),
 	analytics = require( 'analytics' ),
